Rate-limiter: rename to rate-limit.
The module name was incorrect, this fixes that.
