+++ Cargo.toml
@@ -39,7 +39,7 @@ coyote-msgs.workspace = true
 coyote-namespace.workspace = true
 coyote-operations.workspace = true
 coyote-proto.workspace = true
-coyote-rate-limiter.workspace = true
+coyote-rate-limit.workspace = true
 fjall.workspace = true
 fjall-utils.workspace = true
 fs-err.workspace = true
@@ -141,7 +141,7 @@ coyote-msgs.path = "crates/msgs"
 coyote-namespace.path = "crates/coyote-namespace"
 coyote-operations.path = "crates/coyote-operations"
 coyote-proto.path = "crates/coyote-proto"
-coyote-rate-limiter.path = "crates/rate-limiter"
+coyote-rate-limit.path = "crates/rate-limit"
 criterion = "0.8.2"
 ctor = "0.6.3"
 dirs = "6.0.0"
