+++ src/v1/endpoints/rate_limit.rs
@@ -10,7 +10,7 @@ use coyote_derive::aide_annotate;
 use coyote_error::ResultExt;
 use coyote_namespace::Namespace;
 use coyote_proto::MsgPackOrJson;
-use coyote_rate_limiter::RateLimiter;
+use coyote_rate_limit::RateLimiter;
 use jiff::Timestamp;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -19,7 +19,7 @@ use validator::Validate;
 use crate::{AppState, core::cluster::RaftState, error::Result, v1::utils::openapi_tag};
 
 // Re-export types that are used in AppState
-use coyote_rate_limiter::{FixedWindow, RateLimitConfig, RateLimitStatus, TokenBucket};
+use coyote_rate_limit::{FixedWindow, RateLimitConfig, RateLimitStatus, TokenBucket};
 
 pub type RateLimitNamespace = Namespace<RateLimitConfig>;
