+++ src/v1/modules/rate_limit.rs
@@ -1,10 +1,10 @@
 use crate::{AppState, error::Result};
 
-pub use coyote_rate_limiter::RateLimiter;
+pub use coyote_rate_limit::RateLimiter;
 
 /// This is the worker function for this module, it does background cleanup and accounting.
 pub async fn worker(state: AppState) -> Result<()> {
     let stores = [&state.rate_limiter];
-    coyote_rate_limiter::worker(&stores, crate::is_shutting_down).await;
+    coyote_rate_limit::worker(&stores, crate::is_shutting_down).await;
     Ok(())
 }
